+++ crates/plugins/tiangong-plugin-memory/src/lib.rs
@@ -18,6 +18,7 @@
 
 pub mod handler;
 pub mod plugin;
+mod turn_extract;
 
 pub use plugin::MemoryPlugin;
