+++ crates/tiangong-core/src/memory/turn_result.rs
@@ -1,275 +1,10 @@
-//! Memory Turn 记录构建：会话轮次结果提取与持久化
+//! Memory Turn 媒体产物解析：工具结果中的图片/视频资源提取与归档。
 //!
-//! 负责从会话消息中提取工具调用、媒体产物、文件产物等 TurnArtifact，
-//! 构建 TurnResult 供 Memory Actor 写入长期记忆。
-//!
-//! 同时提供工具结果的媒体解析、图片归档等辅助能力，
+//! 记忆轮次提取（候选评估、产物提取、摘要构建）已迁移到 memory 插件的
+//! `turn_extract` 模块。本模块仅保留工具结果的媒体解析与图片归档能力，
 //! 供 ReactEngine 在工具执行完毕后调用。
 
-use std::collections::HashSet;
-
-use crate::session::{Message, MessageRole};
-
-// ── Turn 记忆记录函数 ──
-
-/// 从消息列表中提取所有去重的工具调用名称。
-///
-/// 仅扫描 `Tool` 角色的消息，通过 `parse_tool_calls_line` 和 `parse_tool_trace_name`
-/// 两种格式提取工具名，自动过滤 `recall_memory`，返回去重后的名称列表。
-fn extract_turn_tool_calls(messages: &[Message]) -> Vec<String> {
-    let mut seen = HashSet::new();
-    let mut names = Vec::new();
-    for message in messages {
-        if message.role != MessageRole::Tool {
-            continue;
-        }
-        let text = message.text_content();
-        for name in parse_tool_calls_line(&text)
-            .into_iter()
-            .chain(parse_tool_trace_name(&text))
-        {
-            if name == "recall_memory" {
-                continue;
-            }
-            if seen.insert(name.clone()) {
-                names.push(name);
-            }
-        }
-    }
-    names
-}
-
-/// 从消息列表中提取所有媒体、文件、工具结果类型的 TurnArtifact。
-///
-/// 扫描 `Assistant` 消息中的 media 字段和 `Tool` 消息中的工具追踪内容，
-/// 提取图片/视频 URL、写入文件路径、工具结果摘要等产物，
-/// 通过 URL/path 去重，最多返回 12 个产物。
-fn extract_turn_artifacts(messages: &[Message]) -> Vec<tiangong_memory::TurnArtifact> {
-    let mut artifacts = Vec::new();
-    let mut seen = HashSet::new();
-    for message in messages {
-        if message.role == MessageRole::Assistant {
-            for block in &message.content {
-                if let crate::session::ContentBlock::Media {
-                    kind: _,
-                    url,
-                    title,
-                    ..
-                } = block
-                {
-                    let key = format!("media:{url}");
-                    if seen.insert(key) {
-                        artifacts.push(tiangong_memory::TurnArtifact {
-                            kind: tiangong_memory::TurnArtifactKind::Media,
-                            tool_name: None,
-                            title: title.clone(),
-                            url: Some(url.clone()),
-                            path: None,
-                            summary: None,
-                        });
-                    }
-                }
-            }
-            continue;
-        }
-
-        if message.role != MessageRole::Tool {
-            continue;
-        }
-        let text = message.text_content();
-        let tool_name = parse_tool_trace_name(&text);
-        for artifact in parse_media_artifacts_from_tool_trace(&text, tool_name.as_deref()) {
-            let key = artifact
-                .url
-                .as_deref()
-                .or(artifact.path.as_deref())
-                .unwrap_or_default()
-                .to_string();
-            if !key.is_empty() && seen.insert(key) {
-                artifacts.push(artifact);
-            }
-        }
-        if let Some(path) = tool_name
-            .as_deref()
-            .filter(|name| *name == "write_file" || *name == "replace_in_file")
-            .and_then(|_| parse_written_path(&text))
-        {
-            let key = format!("file:{path}");
-            if seen.insert(key) {
-                artifacts.push(tiangong_memory::TurnArtifact {
-                    kind: tiangong_memory::TurnArtifactKind::File,
-                    tool_name: tool_name.clone(),
-                    title: Some("文件产物".to_string()),
-                    url: None,
-                    path: Some(path),
-                    summary: parse_summary_line(&text),
-                });
-            }
-        }
-        if let Some(tool_name) = tool_name
-            && should_record_tool_result(&tool_name)
-        {
-            let summary =
-                parse_summary_line(&text).unwrap_or_else(|| compact_single_memory_text(&text, 240));
-            let key = format!("tool:{tool_name}:{summary}");
-            if !summary.is_empty() && seen.insert(key) {
-                artifacts.push(tiangong_memory::TurnArtifact {
-                    kind: tiangong_memory::TurnArtifactKind::ToolResult,
-                    tool_name: Some(tool_name),
-                    title: Some("工具结果".to_string()),
-                    url: None,
-                    path: None,
-                    summary: Some(summary),
-                });
-            }
-        }
-    }
-    artifacts.into_iter().take(12).collect()
-}
-
-/// 构建轮次摘要文本。
-///
-/// 优先取最后一条非空 Assistant 消息的压缩内容作为摘要；
-/// 如果没有 Assistant 消息，则将所有产物的 URL/path/summary 拼接为摘要。
-fn build_turn_memory_summary(
-    messages: &[Message],
-    artifacts: &[tiangong_memory::TurnArtifact],
-) -> String {
-    let assistant_summary = messages
-        .iter()
-        .rev()
-        .find(|message| {
-            message.role == MessageRole::Assistant && !message.text_content().trim().is_empty()
-        })
-        .map(|message| compact_single_memory_text(&message.text_content(), 600))
-        .unwrap_or_default();
-    if !assistant_summary.is_empty() {
-        return assistant_summary;
-    }
-    artifacts
-        .iter()
-        .filter_map(|artifact| {
-            artifact
-                .url
-                .as_deref()
-                .or(artifact.path.as_deref())
-                .or(artifact.summary.as_deref())
-                .map(str::to_string)
-        })
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-// ── 解析辅助函数 ──
-
-/// 解析 `tool_calls:` 前缀行，提取逗号分隔的工具调用名称列表。
-///
-/// 例如 `tool_calls: read_file, write_file` 返回 `["read_file", "write_file"]`。
-/// 如果不存在该前缀行，返回空 Vec。
-fn parse_tool_calls_line(content: &str) -> Vec<String> {
-    content
-        .lines()
-        .find_map(|line| line.trim().strip_prefix("tool_calls:"))
-        .map(|raw| {
-            raw.split(',')
-                .map(str::trim)
-                .filter(|name| !name.is_empty())
-                .map(String::from)
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-/// 从工具追踪消息中提取工具名称。
-///
-/// 匹配格式为 `工具执行 [tool_name]` 的首行，返回方括号内的工具名。
-pub(crate) fn parse_tool_trace_name(content: &str) -> Option<String> {
-    let first_line = content.lines().next()?.trim();
-    let rest = first_line.strip_prefix("工具执行 [")?;
-    let end = rest.find(']')?;
-    Some(rest[..end].to_string()).filter(|name| !name.is_empty())
-}
-
-/// 解析 `summary:` 前缀行，提取单行摘要文本。
-///
-/// 返回第一个匹配的 summary 行内容（去除前后空白），
-/// 如果不存在或为空则返回 None。
-fn parse_summary_line(content: &str) -> Option<String> {
-    content.lines().find_map(|line| {
-        line.trim()
-            .strip_prefix("summary:")
-            .map(str::trim)
-            .map(String::from)
-            .filter(|item| !item.is_empty())
-    })
-}
-
-/// 将文本压缩到指定字符数以内。
-///
-/// 先去除空行并合并为单段文本，如果字符数超过 `max_chars` 则截断并追加 `...`。
-/// 用于生成记忆摘要等场景，避免过长文本占用存储空间。
-pub(crate) fn compact_single_memory_text(text: &str, max_chars: usize) -> String {
-    let normalized = text
-        .lines()
-        .map(str::trim)
-        .filter(|line| !line.is_empty())
-        .collect::<Vec<_>>()
-        .join("\n");
-    if normalized.chars().count() <= max_chars {
-        return normalized;
-    }
-    let mut clipped = normalized.chars().take(max_chars).collect::<String>();
-    clipped.push_str("...");
-    clipped
-}
-
 // ── 媒体产物解析函数 ──
-
-/// 从工具追踪消息中解析媒体类型的 TurnArtifact。
-///
-/// 逐行扫描内容，提取 Markdown 图片语法 `![title](url)` 和视频 URL 行，
-/// 将其转换为 `TurnArtifactKind::Media` 类型的产物条目。
-fn parse_media_artifacts_from_tool_trace(
-    content: &str,
-    tool_name: Option<&str>,
-) -> Vec<tiangong_memory::TurnArtifact> {
-    let mut artifacts = Vec::new();
-    let summary = parse_summary_line(content);
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with("![") && trimmed.ends_with(')') {
-            let Some(close_alt) = trimmed.find("](") else {
-                continue;
-            };
-            let title = trimmed[2..close_alt].trim();
-            let url = trimmed[close_alt + 2..trimmed.len() - 1].trim();
-            if url.is_empty() {
-                continue;
-            }
-            artifacts.push(tiangong_memory::TurnArtifact {
-                kind: tiangong_memory::TurnArtifactKind::Media,
-                tool_name: tool_name.map(String::from),
-                title: (!title.is_empty()).then(|| title.to_string()),
-                url: Some(url.to_string()),
-                path: None,
-                summary: summary.clone(),
-            });
-            continue;
-        }
-
-        if let Some(url) = parse_video_url_line(trimmed) {
-            artifacts.push(tiangong_memory::TurnArtifact {
-                kind: tiangong_memory::TurnArtifactKind::Media,
-                tool_name: tool_name.map(String::from),
-                title: Some("生成的视频".to_string()),
-                url: Some(url),
-                path: None,
-                summary: summary.clone(),
-            });
-        }
-    }
-    artifacts
-}
 
 /// 从工具结果中解析 MediaAsset 列表（图片或视频）。
 ///
@@ -429,163 +164,4 @@ fn parse_video_url_line(line: &str) -> Option<String> {
         .map(str::trim)?;
     let url = raw.split_whitespace().next().unwrap_or(raw);
     (url.starts_with("http://") || url.starts_with("https://")).then(|| url.to_string())
-}
-
-/// 解析工具追踪消息中的写入文件路径。
-///
-/// 从 `命令:` 前缀行中提取 `path=...` 参数值，
-/// 用于记录文件写入操作的目标路径。
-fn parse_written_path(content: &str) -> Option<String> {
-    let command = content.lines().find_map(|line| {
-        line.trim()
-            .strip_prefix("命令:")
-            .map(str::trim)
-            .filter(|item| !item.is_empty())
-    })?;
-    let rest = command.strip_prefix("path=")?;
-    let end = rest.find(" content=").unwrap_or(rest.len());
-    Some(rest[..end].trim().to_string()).filter(|path| !path.is_empty())
-}
-
-/// 判断工具结果是否应记录为 TurnArtifact。
-///
-/// 只读类工具（`read_file`、`list_dir`、`tree_dir`、`search_code`、
-/// `recall_memory`、`get_skill_detail`）的结果不需要记录，
-/// 其他工具的结果均应记录。
-fn should_record_tool_result(tool_name: &str) -> bool {
-    !matches!(
-        tool_name,
-        "read_file"
-            | "list_dir"
-            | "tree_dir"
-            | "search_code"
-            | "recall_memory"
-            | "get_skill_detail"
-    )
-}
-
-// ── 记忆候选评估 ──
-
-/// 工具执行后生成记忆候选。
-///
-/// 将工具执行结果结构化记录，由 Memory LLM 在增强版反刍中判断
-/// 是否值得记忆以及提取哪些类型（Episode/Entity/Decision/Evidence）。
-pub(crate) fn evaluate_tool_result_for_memory(
-    tool_name: &str,
-    success: bool,
-    result_summary: &str,
-    file_path: Option<&str>,
-    step_index: usize,
-) -> Option<tiangong_memory::MemoryCandidate> {
-    let summary_trimmed = result_summary.trim();
-    if summary_trimmed.is_empty() {
-        return None;
-    }
-
-    Some(tiangong_memory::MemoryCandidate {
-        tool_name: tool_name.to_string(),
-        step_index,
-        hint: String::new(),
-        suggested_kinds: Vec::new(),
-        file_path: file_path.map(String::from),
-        url: None,
-        result_summary: Some(compact_single_memory_text(summary_trimmed, 240)),
-        success,
-    })
-}
-
-/// 构建增强版轮次结果，附加候选列表和对话消息。
-pub(crate) fn build_enhanced_memory_turn_result(
-    session: &crate::session::Session,
-    turn_start_idx: usize,
-    user_input: &str,
-    candidates: Vec<tiangong_memory::MemoryCandidate>,
-) -> tiangong_memory::EnhancedTurnResult {
-    let messages = session.messages.get(turn_start_idx..).unwrap_or_default();
-    let tool_calls = extract_turn_tool_calls(messages);
-    let artifacts = extract_turn_artifacts(messages);
-    let summary = build_turn_memory_summary(messages, &artifacts);
-    let turn_messages = messages
-        .iter()
-        .filter_map(|message| {
-            let role = match message.role {
-                MessageRole::User => "user",
-                MessageRole::Assistant => "assistant",
-                MessageRole::System => return None,
-                MessageRole::Tool => "tool",
-            };
-            let content = compact_single_memory_text(&message.text_content(), 400);
-            (!content.is_empty()).then(|| tiangong_memory::TurnMessage {
-                role: role.to_string(),
-                content,
-            })
-        })
-        .collect();
-
-    tiangong_memory::EnhancedTurnResult {
-        session_id: session.id.clone(),
-        turn_id: scru128::new().to_string(),
-        had_tool_calls: !tool_calls.is_empty(),
-        user_input: user_input.to_string(),
-        summary,
-        tool_calls,
-        artifacts,
-        workspace_id: None,
-        memory_candidates: candidates,
-        turn_messages,
-    }
-}
-
-// ── 插件入口：从 session 构建增强反刍结果（含候选评估）──
-
-/// 从会话消息构建增强版轮次结果，并评估本轮所有工具结果生成记忆候选。
-///
-/// 供 memory 插件在 `on_turn_finished` 钩子中调用：传入 `&Session` + `turn_start_idx`
-/// + 用户输入，返回完整的 `EnhancedTurnResult`（含候选、产物、消息）。
-///
-/// 插件拿到结果后调 `handle.run_enhanced_micro_rumination_blocking(result)`。
-///
-/// 候选评估从 session 的 tool result 消息回溯（替代原 ReactEngine 内部的
-/// `submit_memory_candidate` 逐条提交），确保插件无需感知 engine 内部状态。
-pub fn build_turn_memory_result(
-    session: &crate::session::Session,
-    turn_start_idx: usize,
-    user_input: &str,
-) -> tiangong_memory::EnhancedTurnResult {
-    // 从本轮消息中评估记忆候选（替代原 engine 的逐条 submit_memory_candidate）。
-    let messages = session.messages.get(turn_start_idx..).unwrap_or_default();
-    let mut candidates = Vec::new();
-    let mut step_index = 0usize;
-    // 构建 tool_call_id → path 映射（从 assistant 消息的 tool_calls 参数提取），
-    // 供 write_file/replace_in_file 候选关联文件路径。
-    let mut path_by_call_id: std::collections::HashMap<&str, &str> =
-        std::collections::HashMap::new();
-    for msg in messages {
-        if msg.role == MessageRole::Assistant {
-            for tc in &msg.tool_calls {
-                if let Some(path) = tc.arguments.get("path").and_then(|v| v.as_str()) {
-                    path_by_call_id.insert(tc.id.as_str(), path);
-                }
-            }
-        }
-    }
-    for msg in messages {
-        if msg.role != MessageRole::Tool {
-            continue;
-        }
-        let tool_name = msg.tool_name.as_deref().unwrap_or("");
-        let success = !msg.tool_result_is_error;
-        let summary = msg.text_content();
-        let file_path = path_by_call_id
-            .get(msg.tool_call_id.as_deref().unwrap_or_default())
-            .copied();
-        if let Some(candidate) =
-            evaluate_tool_result_for_memory(tool_name, success, &summary, file_path, step_index)
-        {
-            candidates.push(candidate);
-            step_index += 1;
-        }
-    }
-
-    build_enhanced_memory_turn_result(session, turn_start_idx, user_input, candidates)
 }
